Move retry inside assign state
