+++ dgraphtest/cluster.go
@@ -90,6 +90,22 @@ type LicenseResponse struct {
 	Extensions map[string]interface{} `json:"license,omitempty"`
 }
 
+type AssignError struct {
+	Errors []Entry `json:"errors"`
+}
+
+type Entry struct {
+	Message    string                 `json:"message"`
+	Extensions map[string]interface{} `json:"extensions"`
+}
+
+type AssignIdResponse struct {
+	StartId  string `json:"startId"`
+	EndId    string `json:"endId"`
+	ReadOnly string `json:"readOnly"`
+}
+
+
 func (hc *HTTPClient) Login(user, password string, ns uint64) error {
 	login := `mutation login($userId: String, $password: String, $namespace: Int, $refreshToken: String) {
 		login(userId: $userId, password: $password, namespace: $namespace, refreshToken: $refreshToken) {
@@ -589,11 +605,55 @@ func (hc *HTTPClient) GetZeroState() (*LicenseResponse, error) {
 	return &stateResponse, nil
 }
 
-func (hc *HTTPClient) AssignState(what string, nums int) (*http.Response, error) {
+
+func parseAssignError(body []byte) (AssignError, error) {
+	var resp AssignError
+	err := json.Unmarshal(body, &resp)
+	return resp, err
+}
+
+func parseAssignIdResponse(body []byte) (AssignIdResponse, error) {
+	var resp AssignIdResponse
+	err := json.Unmarshal(body, &resp)
+	return resp, err
+}
+
+func readResponseBody(do *http.Response) error {
+	defer func() { _ = do.Body.Close() }()
+	body, err := io.ReadAll(do.Body)
+	if err != nil {
+		return err
+	}
+
+	if resp, err := parseAssignError(body); err == nil && len(resp.Errors) > 0 {
+		return fmt.Errorf("failed to assign state. error: %s", resp.Errors[0].Message)
+	}
+
+	if resp, err := parseAssignIdResponse(body); err == nil && resp.StartId != "" {
+		log.Printf("[INFO] success in assign state. StartId: %s EndId: %s", resp.StartId, resp.EndId)
+	}
+
+	return nil
+}
+
+func (hc *HTTPClient) AssignState(what string, nums int) error {
 	log.Printf("[INFO] moving state: %s to %d", what, nums)
 	url := hc.assignURL + "?what=" + what + "&num=" + strconv.Itoa(nums)
-	response, err := http.Get(url)
-	return response, err
+	var err error
+	for i:=0; i<5; i++ {
+		response, err := http.Get(url)
+		if err != nil {
+			log.Printf("[WARN] unable to assing state. err: %s", err)
+			time.Sleep(waitDurBeforeRetry)
+			continue
+		}
+		if err = readResponseBody(response); err == nil {
+			return nil
+		}
+		log.Printf("[WARN] unable to assing state. err: %s", err)
+		time.Sleep(waitDurBeforeRetry)
+	}
+	return err
 }
 
 // SetupSchema sets up DQL schema
